refactor(delegation): drop the in-band attempt ledger from the launch path
`inBandSubagentRun` no longer knows about physical attempts: it launches the
run id its caller derived and reads the child's own record back.

- Deleted `executeStableSubagentInBand`, `StableInBandSubagentRunOptions` with
  its `onActiveRunId` callback, and the module-level `stableRuns` semaphore map
  (the checkpoint lane the whole invocation runs under already serializes two
  dispatches of one call, and the call key derives from that checkpoint).
- Deleted every `stableAttempt` branch in `executeInBand`: the `launched`
  marker in the launch guard, the post-drain `committed` attestation with its
  `afterArtifactsDrained` hook, and every `throwRetryableDurabilityError`. The
  child's `run.start`, `run.result` and `run.end` rows are the launch and
  completion facts; the parent copy was a second at-most-once mechanism.
- Deleted the `mode === 'required-result'` read-back verification of the result
  manifest and the two post-completion loop-failure branches. A loop failure
  after the turn settled (lease release, artifact cleanup) no longer rewrites a
  committed outcome, and a child that failed now throws its own error instead
  of a durability wrapper around it.
- `SubagentDurabilityError` moves here from `stableSubagentAttempt`, the file
  step 4 deletes. It stays the one named type separating an infrastructure
  fault from the child's own failure, so `workflowRunnerError` still maps it to
  `WorkflowRunAbortError` and a durability fault aborts the workflow instead of
  returning null to the script.
- Rewrote the module doc comment, which advertised the attempt ledger, and the
  `childRunLoop` comment that named `stableSubagentAttempt` as the reader of
  the settled `child.turn` row.
- `DelegationHeadless.vitest.ts` drives the surviving entry point;
  `WorkflowScriptAgentRunner.vitest.ts` keeps the real error class through its
  in-band mock. Fifteen attempt-ledger cases in DelegationHeadless now fail as
  expected and are step 5's work.
