+++ src/agent/runtime/childRunLoop.ts
@@ -661,9 +661,9 @@ const deliverTurn = Effect.fn('childRunLoop.deliverTurn')(function* <
     persistChildRunDelivery(params.session, runId, msg, resultMeta),
   );
   // The turn settled whatever the delivery persistence did: its settle path
-  // ran, which is the fact a durable caller's re-execution gate reads
-  // (`stableSubagentAttempt`), so the row lands before that failure is
-  // raised.
+  // ran, which is the fact a recovering caller's re-execution gate reads (a
+  // settled `child.turn` under a run with no outcome refuses repetition), so
+  // the row lands before that failure is raised.
   yield* commitChildTurn(params.session, runId, turnKey, 'settled');
 
   params.onTurnSettled?.({
